Address issue #12 "Why not a minor mode?"
This commit re-implements 'mu4e-dashboard' as a minor mode, removing
the need for the variable 'mu4e-dashboard--buffer' and allowing
multiple dashboards to be active at the same time.

Instead of "activating" a dashboard, one just turns on the minor
mode. This will install a *copy* of the local keymap, updated with
any keybindings defined in the current dashboard. In this way,
multiple dashboards can be active simultaneously, each with their
own local keymap.